backend/retail_market/requirements.txt
ansicon==1.89.0
arrow==1.2.2
asgiref==3.4.1
auth0-python==3.16.2
blessed==1.19.1
boto3==1.18.32
botocore==1.21.32
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
Django==3.2.19
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-health-check==3.16.4
django-pglocks==1.0.4
django-picklefield==3.0.1
django-q==1.3.9
django-q-sentry==0.1.6
django-redis==5.0.0
django-rest-enumfield==0.2.0
django-searchable-encrypted-fields==0.1.9
django-simple-history==3.0.0
django-storages==1.12.3
djangorestframework==3.12.4
drf-jwt==1.19.2
factory-boy==3.2.0
Faker==8.13.2
future==0.18.3
idna==3.2
jinxed==1.1.0
jmespath==0.10.0
nanoid==2.0.0
Pillow==9.3.0
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.14.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-http-client==3.3.2
python-slugify==5.0.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.8.0
sentry-sdk==1.13.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.1
text-unidecode==1.3
urllib3==1.26.11
docusign-esign==3.19.0
wcwidth==0.2.5
channels==3.0.4
daphne==3.0.2
channels-redis==3.3.1
djangorestframework-simplejwt==4.0.0
django-fsm==2.8.0
pyexcel-xlsx==0.6.0
weasyprint==53.4
django-mathfilters==1.0.0
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
textx==3.1.1
tblib==1.7.0